+++ fresh.gen.ts
@@ -17,18 +17,19 @@ import * as $11 from "./routes/blog/index.tsx";
 import * as $12 from "./routes/callback.ts";
 import * as $13 from "./routes/dashboard/_middleware.ts";
 import * as $14 from "./routes/dashboard/index.tsx";
-import * as $15 from "./routes/dashboard/users.tsx";
-import * as $16 from "./routes/feed.ts";
-import * as $17 from "./routes/index.tsx";
-import * as $18 from "./routes/item/[id].tsx";
-import * as $19 from "./routes/notifications/[id].ts";
-import * as $20 from "./routes/notifications/_middleware.ts";
-import * as $21 from "./routes/notifications/index.tsx";
-import * as $22 from "./routes/pricing.tsx";
-import * as $23 from "./routes/signin.ts";
-import * as $24 from "./routes/signout.ts";
-import * as $25 from "./routes/submit.tsx";
-import * as $26 from "./routes/user/[login].tsx";
+import * as $15 from "./routes/dashboard/stats.tsx";
+import * as $16 from "./routes/dashboard/users.tsx";
+import * as $17 from "./routes/feed.ts";
+import * as $18 from "./routes/index.tsx";
+import * as $19 from "./routes/item/[id].tsx";
+import * as $20 from "./routes/notifications/[id].ts";
+import * as $21 from "./routes/notifications/_middleware.ts";
+import * as $22 from "./routes/notifications/index.tsx";
+import * as $23 from "./routes/pricing.tsx";
+import * as $24 from "./routes/signin.ts";
+import * as $25 from "./routes/signout.ts";
+import * as $26 from "./routes/submit.tsx";
+import * as $27 from "./routes/user/[login].tsx";
 import * as $$0 from "./islands/Chart.tsx";
 import * as $$1 from "./islands/PageInput.tsx";
 import * as $$2 from "./islands/VoteButton.tsx";
@@ -50,18 +51,19 @@ const manifest = {
     "./routes/callback.ts": $12,
     "./routes/dashboard/_middleware.ts": $13,
     "./routes/dashboard/index.tsx": $14,
-    "./routes/dashboard/users.tsx": $15,
-    "./routes/feed.ts": $16,
-    "./routes/index.tsx": $17,
-    "./routes/item/[id].tsx": $18,
-    "./routes/notifications/[id].ts": $19,
-    "./routes/notifications/_middleware.ts": $20,
-    "./routes/notifications/index.tsx": $21,
-    "./routes/pricing.tsx": $22,
-    "./routes/signin.ts": $23,
-    "./routes/signout.ts": $24,
-    "./routes/submit.tsx": $25,
-    "./routes/user/[login].tsx": $26,
+    "./routes/dashboard/stats.tsx": $15,
+    "./routes/dashboard/users.tsx": $16,
+    "./routes/feed.ts": $17,
+    "./routes/index.tsx": $18,
+    "./routes/item/[id].tsx": $19,
+    "./routes/notifications/[id].ts": $20,
+    "./routes/notifications/_middleware.ts": $21,
+    "./routes/notifications/index.tsx": $22,
+    "./routes/pricing.tsx": $23,
+    "./routes/signin.ts": $24,
+    "./routes/signout.ts": $25,
+    "./routes/submit.tsx": $26,
+    "./routes/user/[login].tsx": $27,
   },
   islands: {
     "./islands/Chart.tsx": $$0,
